task: moving hctx to the render package
