+++ internal/helpers/content/content.go
@@ -1,6 +1,6 @@
 package content
 
-import "github.com/leapkit/core/hctx"
+import "github.com/leapkit/core/render/hctx"
 
 // Keys to be used in templates for the functions in this package.
 const (
