+++ internal/helpers/content/for_test.go
@@ -4,8 +4,8 @@ import (
 	"errors"
 	"testing"
 
-	"github.com/leapkit/core/hctx"
 	"github.com/leapkit/core/internal/helpers/helptest"
+	"github.com/leapkit/core/render/hctx"
 	"github.com/stretchr/testify/require"
 )
