+++ internal/helpers/debug/debug.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"html/template"
 
-	"github.com/leapkit/core/hctx"
+	"github.com/leapkit/core/render/hctx"
 )
 
 // Keys to be used in templates for the functions in this package.
