+++ internal/helpers/meta/meta.go
@@ -1,7 +1,7 @@
 package meta
 
 import (
-	"github.com/leapkit/core/hctx"
+	"github.com/leapkit/core/render/hctx"
 )
 
 // Keys to be used in templates for the functions in this package.
