+++ render/hctx/map.go
